Examen ExtraOrd Programacion ayuda

+++ Servicios y Procesos/ExamenExtraOrd/src/main/java/Ejer_3/Interfaces/NutricionCalorias.java
@@ -7,5 +7,5 @@ public interface NutricionCalorias {
     double EMPERADOR = 300;
     double SUSHI = 100;
 
-    int digerir();
+    double digerir();
 }
